add toggle button and refacto components
